requirements.txt
APScheduler==3.7.0
bidict==0.21.2
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
country-converter==0.7.2
dnspython==2.1.0
dotmap==1.3.23
email-validator==1.1.2
Flask==1.1.2
Flask-APScheduler==1.12.1
Flask-Login==0.5.0
Flask-Navigation==0.2.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flickrapi==2.4.0
geographiclib==1.50
geopy==2.1.0
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.6.2
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
plotly==4.14.3
psycopg2-binary==2.8.6
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.4.3
textblob==0.15.3
tqdm==4.60.0
tweepy==3.10.0
tzlocal==2.1
urllib3==1.26.4
validate-email==1.3
Werkzeug==1.0.1
WTForms==2.3.3
